requirements.txt
fastapi
uvicorn
langchain-google-genai
langchain-community
tavily-python
langchain-mongodb
python-dotenv
pydantic